9043/refactor: extract order related methods from UBSClientService to OrderService
